Move active series to ingester package.
